spiders: add base class for parsing
